Fix errors/warnings in Chrome
Use webextension-polyfill fork to fix
https://github.com/mozilla/webextension-polyfill/issues/130

Make sure no `onMessage` listeners are `async`. `WorkerProgram` and
`RendererProgram` both listen to all messages from background, but only
handle the ones meant for them based on an `if` statement. However, if
the listener is an `async` function it will *always* return a Promise,
indicating that the listener wants to send a response. It is only
allowed to send *one* response per message per document. Otherwise
Chrome shows an error.