Fix frame timeout logic
Previously, if there was a frame we accidentally always waited the 100ms
timeout even if the frame was “alive”. This commit keeps track of how
many frames we are waiting for to see if they are alive or not. If that
number goes to zero, clear the timeout and allow entering hinting as
soon as possible.