Handle elements with several icons, some hidden
The navigation tabs at the top of Twitter have _two_ icons in them – one
outline variant, and one filled variant. Twitter toggles which one is
shown based on if the tab is active or not. Previously, we only tried
placing the hint at the first icon in DOM order before giving up. This
commit looks at several icons in DOM order, and tries to use the first
one that is not hidden. One _could_ try to find the left-most of the
icons, but so far I don't think it's worth the trouble anywhere.