Add missing `break`
There used to be a `return` here, which correctly stopped the loop when
the deadline was passed. Then, the `timesUp` flag was added because a
little bit of code _after_ the loop needs to be run as well. However
that commit forgot to remove the `return` it was replacing. That was
later done in commit 61c0c198. But what really should have been done was
replacing the `return` with a `break` – we want to stop the loop _and_
run some code after it.