Fix performance issues on pages with lots of nodes
- Do as little as possible in `onMutation`. Move the work to `flushQueue`.
- Similarly, do as little in event listeners as well.
- Don't process the same elements more than once in the queue. It turns
out that MutationObservers are triggered for _every_ new element (not
just for inserted parents) when large HTML responses are streamed in
chunks. Since we do `.querySelectorAll("*")` on all added elements, it
meant that we processed children many times.

The above means that:

- The queue flushes faster after page load.
- If entering hints mode very quickly at page load, the synchronous
`flushQueue` is now less likely to trigger the "an extension script is
slowing your browser down" message.
- The background monitoring of the page is now even less likely to
affect the page performance, since minimal work is done immediately and
moved to the `requestIdleCallback`-powered `flushQueue`.