Improve hinting in gmail
There’s a div with a click listener around the table with starred email.
That div unfortunately doesn’t do anything when clicked, but I haven’t
found a way of getting rid of its hint.

Previously, that hint got a weird position. If there was just one
starred email and that email had an attachment, the hint ended up at the
attachment, slightly off vertically.

It was slightly off due to hidden screen reader only text. This commit
fixes the screen reader text detection.

Then the hint moved to the text of the attachments. But why the
attachments? Why not some other text earlier in the email row?

All the text in the email row has the same height. Previously, in case
of a tie the _widest_ element was chosen, which was the attachment text
in this case. This commit changes to the _left-most_ text instead.

Moving that hint is important because it made it look like you could
click the attachments (which unfortunately aren’t supported).