Improve hint positions on dietdoctor.com
https://www.dietdoctor.com/low-carb/recipes/meals

At the top there’s a horizontal bar of recipe categories. Each category
is a link with an image and some text below it. Sometimes the text is
wider than the image, sometimes not. When the text is wider, Link Hints
would place the hint at the text, but when the image is wider, the hint
would end up at the image. This felt pretty inconsistent.

This happened because of the logic to prefer to place hints to the left
of an icon just to the left of text. However, that logic had forgotten
to check if the icon is also on the same line.