OpenNewTab: Support Vivaldi
Vivaldi is a Chromium-based browser that has decided to present itself
as Chrome but that differs from Chrome in a number of ways. In
particular, Vivaldi uses shift+click to open a link in a new foreground
tab and ctrl+shift+click to open a link in a new foreground *window*.

This adds support to detect Vivaldi and change the simulated click that
the "OpenNewTab" command executes accordingly. Per this SO answer[1], we
must use the extension tabs API to identify Vivaldi, and this requires
some extra work that we'd prefer to do just once, so we do it as part of
the options setup on the background Program. Additionally, the check
requires accessing a property that the WebExtension definition of
`browser.tabs.Tab` doesn't know about, so we have to extend that
interface in @types.globals.d.ts for this special case.

[1]: https://stackoverflow.com/a/77047611

Closes https://github.com/lydell/LinkHints/issues/39