Avoid placing hints a `<sup>` elements
Closes #84. In that issue, it looked like some sentences got no hints.
In reality, they did, but they were placed at `<sup>` elements in the
sentences and ended up behind the hints for selecting only the `<sup>`
text.

This PR tries to detect potential `<sup>` elements and prefers the start
of the text instead in those cases.

Unfortunately it is very difficult to tell `<sup>` and wrapped text
apart, so this isn't perfect. But at least it's better. Probably good in
all realistic cases.