Remove unecessary, complicated logic that resulted in a false positive
Closes #73.

The removed code comment explained the edge cases it was for. We have
test cases for them in html/hint-position/index.html and
html/labels/index.html. For the edge case with tabs, it still worked
without the extra code. For the edge case with the Bootstrap v4
checkbox, the hint is now at the checkbox label instead of at the fake
checkbox, which is totally ok.

What happened in the issue was that GitHub used `visibility: hidden` on
a dropdown menu. When an item inside the invisible menu aligned
perfectly on top of a visible element on the page with a certain size,
that special case code resulted in the invisible menu item being treated
as visible after all.