Fix order of mousedown, focus and mouseup
We used a wrong order of mousedown, focus and mouseup (compared to what
can be observed when making a real click) on purpose as a workaround for
Gmail. This caused some dropdowns not to work on Codeberg. By switching
to the correct event order, Codeberg works. I’ve tested Gmail, and
didn’t notice any problems anymore.

Closes #76.