Remove IntersectionObserver
I’ve noticed that tracking all clickable elements with
`IntersectionObserver` does not seem to speed things up noticeably.
Instead, it just runs the risk of slowing down the browser. This commit
removes the `IntersectionObserver`. We had support for it not running
anyway, since it makes scrolling laggy on massive pages – now we use
that approach everywhere.

Not that we still use `IntersectionObserver` for frames, and in the
renderer.